+++ spec/www/spec/db-tx-error-mapping-test.js
@@ -8,9 +8,10 @@ var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
 var isWindows = /Windows /.test(navigator.userAgent); // Windows
 var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
-// NOTE: In the common storage-master branch there is no difference between the
-// default implementation and implementation #2. But the test will also apply
-// the androidLockWorkaround: 1 option in the case of implementation #2.
+// NOTE: While in certain version branches there is no difference between
+// the default Android implementation and implementation #2,
+// this test script will also apply the androidLockWorkaround: 1 option
+// in case of implementation #2.
 var scenarioList = [
   isAndroid ? 'Plugin-implementation-default' : 'Plugin',
   'HTML5',
@@ -105,18 +106,23 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 near \"SLCT\": syntax error/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/syntax error or other error.*code 1/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"SLCT\": syntax error.*code 1.*while compiling: SLCT 1/);
               else
                 expect(error.message).toMatch(/near \"SLCT\": syntax error/);
@@ -130,6 +136,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || isWebSql || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -151,7 +162,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT syntax error [VALUES in the wrong place (XXX "incomplete input")] with a trailing space [XXX TBD PROPER INFO MESSAGE NOT ON Android/...]', function(done) {
+        it(suiteName + 'INSERT syntax error [VALUES in the wrong place] with a trailing space [XXX "incomplete input" message on XXX, XXX TODO PROPER INFO MESSAGE NOT ON Android/XXX]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("INSERT-Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -175,21 +186,31 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 near \"VALUES\": syntax error/);
+              else if (isWebSql)
+                expect(error.message).toMatch(/near \"VALUES\": syntax error/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
+              //* else if (isAndroid && !isImpl2) //* XXX TBD Android (default implementation) vs ...
+              //*   expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*near \" \": syntax error/);
               else if (isAndroid && !isImpl2) //* XXX TBD PROPER INFO MESSAGE NOT ON Android/...
                 expect(error.message).toMatch(/syntax error or other error.*code 1/);
               else if (isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"VALUES\": syntax error.*code 1.*while compiling: INSERT INTO test_table/);
               else
-                expect(error.message).toMatch(/incomplete input/); // XXX SQLite 3.22.0 on iOS/macOS
+                //* expect(error.message).toMatch(/near \" \": syntax error/);
+                expect(error.message).toMatch(/incomplete input/); // XXX SQLite 3.22.0
 
               // FAIL transaction & check reported transaction error:
               return true;
@@ -199,6 +220,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || isWebSql || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -253,18 +279,25 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
-              if (isWebSql && !isAndroid)
-                expect(true).toBe(true); // SKIP for iOS (WebKit) Web SQL
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
+              if (isWebSql && (!isAndroid || /Android 4.[1-3]/.test(navigator.userAgent)))
+                expect(true).toBe(true); // SKIP for iOS (WebKit) & Android 4.1-4.3 (WebKit) Web SQL
               else if (isWindows)
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(6);
 
               // (WebKit) Web SQL (Android/iOS) possibly with a missing 'r'
-              if (isWebSql && isAndroid)
+              if (isWebSql && /Android 4.[1-3]/.test(navigator.userAgent))
+                expect(error.message).toMatch(/column data is not unique/);
+              else if (isWebSql && isAndroid)
                 expect(error.message).toMatch(/could not execute statement due to a constr?aint failure.*19.*constraint failed/);
               else if (isWebSql)
-                expect(error.message).toMatch(/constr?aint fail/);
+                expect(error.message).toMatch(/constr?aint fail/); // [possibly missing letter on iOS (WebKit) Web SQL]
               else if (isWindows)
                 expect(error.message).toMatch(/SQLite3 step error result code: 1/);
               else if (isAndroid && !isImpl2)
@@ -283,6 +316,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || isWebSql)
               expect(error.code).toBe(0);
@@ -325,19 +363,24 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
               // ACTUAL WebKit Web SQL vs plugin error.message
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 no such function: uper/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/syntax error or other error.*code 1/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such function: uper.*code 1/);
               else
                 expect(error.message).toMatch(/no such function: uper/);
@@ -351,6 +394,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -396,18 +444,23 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 no such table: BogusTable/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/syntax error or other error.*code 1/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such table: BogusTable.*code 1/);
               else
                 expect(error.message).toMatch(/no such table: BogusTable/);
@@ -420,6 +473,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -466,18 +524,23 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 table test_table has 2 columns but 1 values were supplied/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/syntax error or other error.*code 1/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/table test_table has 2 columns but 1 values were supplied.*code 1.*while compiling: INSERT INTO test_table/);
               else
                 expect(error.message).toMatch(/table test_table has 2 columns but 1 values were supplied/);
@@ -491,6 +554,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -537,18 +605,23 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 table test_table has no column named wrong_column/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/syntax error or other error.*code 1/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/table test_table has no column named wrong_column.*code 1.*while compiling: INSERT INTO test_table/);
               else
                 expect(error.message).toMatch(/table test_table has no column named wrong_column/);
@@ -562,6 +635,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -610,14 +688,21 @@ var mytests = function() {
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
+              // error.hasOwnProperty('message') apparently NOT WORKING on
+              // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+                expect(error.hasOwnProperty('message')).toBe(true);
+
               if (isWindows || (!isWebSql && isAndroid && isImpl2))
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql && isAndroid)
+              if (isWebSql && isAndroid && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*not authorized/);
-              else if (isWebSql) // (iOS)
+              else if (isWebSql && isAndroid)
+                expect(error.message).toMatch(/not authorized/);
+              else if (isWebSql) // [iOS (WebKit) Web SQL]
                 expect(error.message).toMatch(/could not prepare statement.*1 not authorized/);
               else if (isWindows)
                 expect(error.message).toMatch(/SQLite3 step error result code: 1/);
@@ -636,6 +721,11 @@ var mytests = function() {
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
+
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
@@ -673,21 +763,28 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
+
             if (isWindows || (isAndroid && isImpl2))
               expect(error.code).toBe(0);
             else
               expect(error.code).toBe(5);
 
-            if (isWebSql)
+            if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
               expect(error.message).toMatch(/could not prepare statement.*1 near \"SLCT\": syntax error/);
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
-            else if (isAndroid && !isImpl2)
+            else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
-            else if (isAndroid && isImpl2)
+            else if (!isWebSql && !isWindows && isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: near \"SLCT\": syntax error.*code 1.*while compiling: SLCT 1/);
-            else
+            else if (!isWebSql) // [iOS/macOS plugin]
               expect(error.message).toMatch(/a statement with no error handler failed.*near \"SLCT\": syntax error/);
+            else
+              expect(error.message).toMatch(/near \"SLCT\": syntax error/);
 
             isWebSql ? done() : db.close(done, done);
           }, function() {
@@ -715,15 +812,22 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWebSql && !isAndroid)
-              expect(true).toBe(true); // SKIP for iOS (WebKit) Web SQL
+            // error.hasOwnProperty('message') apparently NOT WORKING on
+            // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
+              expect(error.hasOwnProperty('message')).toBe(true);
+
+            if (isWebSql && (!isAndroid || /Android 4.[1-3]/.test(navigator.userAgent)))
+              expect(true).toBe(true); // SKIP for iOS (WebKit) & Android 4.1-4.3 (WebKit) Web SQL
             else if (isWindows)
               expect(error.code).toBe(0);
             else
               expect(error.code).toBe(6);
 
             // (WebKit) Web SQL (Android/iOS) possibly with a missing 'r'
-            if (isWebSql && isAndroid)
+            if (isWebSql && /Android 4.[1-3]/.test(navigator.userAgent))
+              expect(error.message).toMatch(/column data is not unique/);
+            else if (isWebSql && isAndroid)
               expect(error.message).toMatch(/could not execute statement due to a constr?aint failure.*19.*constraint failed/);
             else if (isWebSql)
               expect(error.message).toMatch(/constr?aint fail/);